% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLEXanalyzer-package.R
\docType{data}
\name{ER_ARID1A_KO_MCF7}
\alias{ER_ARID1A_KO_MCF7}
\title{ER_ARID1A_KO_MCF7 dataset}
\format{
An object of class \code{\link{list}} related to peptides
quantification. It consists of qPLEX-RIME data from five experimental runs.
Each run contains 9 samples divided into nine conditions (T_14, V_14,
T_11, V_11, ECACC.T, ECACC.V, T_221, V_221 and Ref).
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
Five ER qPLEX-RIME (9plex) experiments were performed on two wild type clones,
two ARID1A knockout clones and one parental cell line with Tamoxifen treatment
in MCF7 cell lines.
}
\keyword{data}
\keyword{datasets}
