\docType{data}
\name{exp2_Xlink}
\alias{exp2_Xlink}
\title{exp2_Xlink dataset}
\format{An object of class \code{\link{list}} related to peptides
 quantification. It consists of qPLEX-RIME data of 10 samples
 divided into three conditions (FA, DSG.FA and IgG).}
\usage{
data(exp2_Xlink)
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
An ER qPLEX-RIME experiment was performed to compare two different methods of crosslinking. 
MCF7 cells were double crosslinked with DSG/formaldehyde (double) or with formaldehyde alone 
(single). Four biological replicates were obtained for each condition along with two IgG pooled 
samples from each replicate.
}
\keyword{data}
\keyword{datasets}

