% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLEXanalyzer-package.R
\docType{data}
\name{exp3_OHT_ESR1}
\alias{exp3_OHT_ESR1}
\title{exp3_OHT_ESR1 dataset}
\format{
An object of class \code{\link{list}} related to peptides
quantification. It consists of qPLEX-RIME data from three experimental runs.
Each run contains 10 samples divided into five conditions (IgG, vehicle,
tam.2h, tam.6h and tam.24h).
}
\value{
An object of class \code{\link{list}} related to peptides
quantification.
}
\description{
Three ER qPLEX-RIME (10plex) experiments were performed to investigate the
dynamics of the ER complex assembly upon 4-hydrotamoxifen (OHT) treatment at
2h, 6h and 24h or at 24h post-treatment with the drug-vehicle alone
(ethanol). Two biological replicates of each condition were included in each
experiment to finally consider a total of six replicates per time point.
Additionally, MCF7 cells were treated with OHT or ethanol and cross-linked
at 24h post-treatment in each experiment to be used for mock IgG pull-downs
and to enable discrimination of non-specific binding in the same experiment.
This is a timecourse experiment to study the effect of tamoxifen in ER
interactome using qPLEX-RIME method.
}
\keyword{data}
\keyword{datasets}
