\docType{data}
\name{human_anno}
\alias{human_anno}
\title{human_anno dataset}
\description{
  Uniprot Human protein annotation table.
}
\format{An object of class \code{\link{data.frame}} consisting of
uniprot human protein annotation. }
\usage{
data(human_anno)
}
\keyword{data}
\keyword{datasets}

