% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLEXanalyzer-package.R
\docType{data}
\name{mouse_anno}
\alias{mouse_anno}
\title{mouse_anno dataset}
\format{
An object of class \code{\link{data.frame}} consisting of uniprot
mouse protein annotation.
}
\description{
Uniprot Mouse protein annotation table.
}
\keyword{data}
\keyword{datasets}
