% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPLEXanalyzer-package.R
\docType{data}
\name{human_anno}
\alias{human_anno}
\title{human_anno dataset}
\format{
An object of class \code{\link{data.frame}} consisting of uniprot
human protein annotation.
}
\description{
Uniprot Human protein annotation table.
}
\keyword{data}
\keyword{datasets}
